mem.Info memory fields int64 -> uint64

+++ mem/flat/mem_info.go
@@ -75,50 +75,50 @@ func (prof *Profiler) Serialize(inf *mem.Info) []byte {
 	// ensure the Builder is in a usable state.
 	prof.Builder.Reset()
 	InfoStart(prof.Builder)
-	InfoAddTimestamp(prof.Builder, int64(inf.Timestamp))
-	InfoAddActive(prof.Builder, int64(inf.Active))
-	InfoAddActiveAnon(prof.Builder, int64(inf.ActiveAnon))
-	InfoAddActiveFile(prof.Builder, int64(inf.ActiveFile))
-	InfoAddAnonHugePages(prof.Builder, int64(inf.AnonHugePages))
-	InfoAddAnonPages(prof.Builder, int64(inf.AnonPages))
-	InfoAddBounce(prof.Builder, int64(inf.Bounce))
-	InfoAddBuffers(prof.Builder, int64(inf.Buffers))
-	InfoAddCached(prof.Builder, int64(inf.Cached))
-	InfoAddCommitLimit(prof.Builder, int64(inf.CommitLimit))
-	InfoAddCommittedAS(prof.Builder, int64(inf.CommittedAS))
-	InfoAddDirectMap4K(prof.Builder, int64(inf.DirectMap4K))
-	InfoAddDirectMap2M(prof.Builder, int64(inf.DirectMap2M))
-	InfoAddDirty(prof.Builder, int64(inf.Dirty))
-	InfoAddHardwareCorrupted(prof.Builder, int64(inf.HardwareCorrupted))
-	InfoAddHugePagesFree(prof.Builder, int64(inf.HugePagesFree))
-	InfoAddHugePagesRsvd(prof.Builder, int64(inf.HugePagesRsvd))
-	InfoAddHugePagesSize(prof.Builder, int64(inf.HugePagesSize))
-	InfoAddHugePagesSurp(prof.Builder, int64(inf.HugePagesSurp))
-	InfoAddHugePagesTotal(prof.Builder, int64(inf.HugePagesTotal))
-	InfoAddInactive(prof.Builder, int64(inf.Inactive))
-	InfoAddInactiveAnon(prof.Builder, int64(inf.InactiveAnon))
-	InfoAddInactiveFile(prof.Builder, int64(inf.InactiveFile))
-	InfoAddKernelStack(prof.Builder, int64(inf.KernelStack))
-	InfoAddMapped(prof.Builder, int64(inf.Mapped))
-	InfoAddMemAvailable(prof.Builder, int64(inf.MemAvailable))
-	InfoAddMemFree(prof.Builder, int64(inf.MemFree))
-	InfoAddMemTotal(prof.Builder, int64(inf.MemTotal))
-	InfoAddMlocked(prof.Builder, int64(inf.Mlocked))
-	InfoAddNFSUnstable(prof.Builder, int64(inf.NFSUnstable))
-	InfoAddPageTables(prof.Builder, int64(inf.PageTables))
-	InfoAddShmem(prof.Builder, int64(inf.Shmem))
-	InfoAddSlab(prof.Builder, int64(inf.Slab))
-	InfoAddSReclaimable(prof.Builder, int64(inf.SReclaimable))
-	InfoAddSUnreclaim(prof.Builder, int64(inf.SUnreclaim))
-	InfoAddSwapCached(prof.Builder, int64(inf.SwapCached))
-	InfoAddSwapFree(prof.Builder, int64(inf.SwapFree))
-	InfoAddSwapTotal(prof.Builder, int64(inf.SwapTotal))
-	InfoAddUnevictable(prof.Builder, int64(inf.Unevictable))
-	InfoAddVmallocChunk(prof.Builder, int64(inf.VmallocChunk))
-	InfoAddVmallocTotal(prof.Builder, int64(inf.VmallocTotal))
-	InfoAddVmallocUsed(prof.Builder, int64(inf.VmallocUsed))
-	InfoAddWriteback(prof.Builder, int64(inf.Writeback))
-	InfoAddWritebackTmp(prof.Builder, int64(inf.WritebackTmp))
+	InfoAddTimestamp(prof.Builder, inf.Timestamp)
+	InfoAddActive(prof.Builder, inf.Active)
+	InfoAddActiveAnon(prof.Builder, inf.ActiveAnon)
+	InfoAddActiveFile(prof.Builder, inf.ActiveFile)
+	InfoAddAnonHugePages(prof.Builder, inf.AnonHugePages)
+	InfoAddAnonPages(prof.Builder, inf.AnonPages)
+	InfoAddBounce(prof.Builder, inf.Bounce)
+	InfoAddBuffers(prof.Builder, inf.Buffers)
+	InfoAddCached(prof.Builder, inf.Cached)
+	InfoAddCommitLimit(prof.Builder, inf.CommitLimit)
+	InfoAddCommittedAS(prof.Builder, inf.CommittedAS)
+	InfoAddDirectMap4K(prof.Builder, inf.DirectMap4K)
+	InfoAddDirectMap2M(prof.Builder, inf.DirectMap2M)
+	InfoAddDirty(prof.Builder, inf.Dirty)
+	InfoAddHardwareCorrupted(prof.Builder, inf.HardwareCorrupted)
+	InfoAddHugePagesFree(prof.Builder, inf.HugePagesFree)
+	InfoAddHugePagesRsvd(prof.Builder, inf.HugePagesRsvd)
+	InfoAddHugePagesSize(prof.Builder, inf.HugePagesSize)
+	InfoAddHugePagesSurp(prof.Builder, inf.HugePagesSurp)
+	InfoAddHugePagesTotal(prof.Builder, inf.HugePagesTotal)
+	InfoAddInactive(prof.Builder, inf.Inactive)
+	InfoAddInactiveAnon(prof.Builder, inf.InactiveAnon)
+	InfoAddInactiveFile(prof.Builder, inf.InactiveFile)
+	InfoAddKernelStack(prof.Builder, inf.KernelStack)
+	InfoAddMapped(prof.Builder, inf.Mapped)
+	InfoAddMemAvailable(prof.Builder, inf.MemAvailable)
+	InfoAddMemFree(prof.Builder, inf.MemFree)
+	InfoAddMemTotal(prof.Builder, inf.MemTotal)
+	InfoAddMlocked(prof.Builder, inf.Mlocked)
+	InfoAddNFSUnstable(prof.Builder, inf.NFSUnstable)
+	InfoAddPageTables(prof.Builder, inf.PageTables)
+	InfoAddShmem(prof.Builder, inf.Shmem)
+	InfoAddSlab(prof.Builder, inf.Slab)
+	InfoAddSReclaimable(prof.Builder, inf.SReclaimable)
+	InfoAddSUnreclaim(prof.Builder, inf.SUnreclaim)
+	InfoAddSwapCached(prof.Builder, inf.SwapCached)
+	InfoAddSwapFree(prof.Builder, inf.SwapFree)
+	InfoAddSwapTotal(prof.Builder, inf.SwapTotal)
+	InfoAddUnevictable(prof.Builder, inf.Unevictable)
+	InfoAddVmallocChunk(prof.Builder, inf.VmallocChunk)
+	InfoAddVmallocTotal(prof.Builder, inf.VmallocTotal)
+	InfoAddVmallocUsed(prof.Builder, inf.VmallocUsed)
+	InfoAddWriteback(prof.Builder, inf.Writeback)
+	InfoAddWritebackTmp(prof.Builder, inf.WritebackTmp)
 	prof.Builder.Finish(InfoEnd(prof.Builder))
 	p := prof.Builder.Bytes[prof.Builder.Head():]
 	// copy them (otherwise gets lost in reset)
@@ -283,171 +283,171 @@ Tick:
 				if v == 'A' {
 					if t.Val[5] == 'e' {
 						if nameLen == 6 {
-							InfoAddActive(t.Builder, int64(n))
+							InfoAddActive(t.Builder, n)
 							continue
 						}
 						if t.Val[7] == 'a' {
-							InfoAddActiveAnon(t.Builder, int64(n))
+							InfoAddActiveAnon(t.Builder, n)
 							continue
 						}
-						InfoAddActiveFile(t.Builder, int64(n))
+						InfoAddActiveFile(t.Builder, n)
 						continue
 					}
 					if nameLen == 9 {
-						InfoAddAnonPages(t.Builder, int64(n))
+						InfoAddAnonPages(t.Builder, n)
 						continue
 					}
-					InfoAddAnonHugePages(t.Builder, int64(n))
+					InfoAddAnonHugePages(t.Builder, n)
 					continue
 				}
 				if v == 'C' {
 					if nameLen == 6 {
-						InfoAddCached(t.Builder, int64(n))
+						InfoAddCached(t.Builder, n)
 						continue
 					}
 					if nameLen == 11 {
-						InfoAddCommitLimit(t.Builder, int64(n))
+						InfoAddCommitLimit(t.Builder, n)
 						continue
 					}
-					InfoAddCommittedAS(t.Builder, int64(n))
+					InfoAddCommittedAS(t.Builder, n)
 					continue
 				}
 				if v == 'D' {
 					if nameLen == 5 {
-						InfoAddDirty(t.Builder, int64(n))
+						InfoAddDirty(t.Builder, n)
 						continue
 					}
 					if t.Val[10] == 'k' {
-						InfoAddDirectMap4K(t.Builder, int64(n))
+						InfoAddDirectMap4K(t.Builder, n)
 						continue
 					}
-					InfoAddDirectMap2M(t.Builder, int64(n))
+					InfoAddDirectMap2M(t.Builder, n)
 					continue
 				}
 				if v == 'H' {
 					if nameLen == 14 {
 						if t.Val[10] == 'F' {
-							InfoAddHugePagesFree(t.Builder, int64(n))
+							InfoAddHugePagesFree(t.Builder, n)
 							continue
 						}
 						if t.Val[10] == 'R' {
-							InfoAddHugePagesRsvd(t.Builder, int64(n))
+							InfoAddHugePagesRsvd(t.Builder, n)
 							continue
 						}
-						InfoAddHugePagesSurp(t.Builder, int64(n))
+						InfoAddHugePagesSurp(t.Builder, n)
 					}
 					if t.Val[1] == 'a' {
-						InfoAddHardwareCorrupted(t.Builder, int64(n))
+						InfoAddHardwareCorrupted(t.Builder, n)
 						continue
 					}
 					if t.Val[9] == 'i' {
-						InfoAddHugePagesSize(t.Builder, int64(n))
+						InfoAddHugePagesSize(t.Builder, n)
 						continue
 					}
-					InfoAddHugePagesTotal(t.Builder, int64(n))
+					InfoAddHugePagesTotal(t.Builder, n)
 					continue
 				}
 				if v == 'I' {
 					if nameLen == 8 {
-						InfoAddInactive(t.Builder, int64(n))
+						InfoAddInactive(t.Builder, n)
 						continue
 					}
 					if t.Val[9] == 'a' {
-						InfoAddInactiveAnon(t.Builder, int64(n))
+						InfoAddInactiveAnon(t.Builder, n)
 						continue
 					}
-					InfoAddInactiveFile(t.Builder, int64(n))
+					InfoAddInactiveFile(t.Builder, n)
 				}
 				if v == 'M' {
 					v = t.Val[3]
 					if nameLen < 8 {
 						if v == 'p' {
-							InfoAddMapped(t.Builder, int64(n))
+							InfoAddMapped(t.Builder, n)
 							continue
 						}
 						if v == 'F' {
-							InfoAddMemFree(t.Builder, int64(n))
+							InfoAddMemFree(t.Builder, n)
 							continue
 						}
-						InfoAddMlocked(t.Builder, int64(n))
+						InfoAddMlocked(t.Builder, n)
 						continue
 					}
 					if v == 'A' {
-						InfoAddMemAvailable(t.Builder, int64(n))
+						InfoAddMemAvailable(t.Builder, n)
 						continue
 					}
-					InfoAddMemTotal(t.Builder, int64(n))
+					InfoAddMemTotal(t.Builder, n)
 					continue
 				}
 				if v == 'S' {
 					v = t.Val[1]
 					if v == 'w' {
 						if t.Val[4] == 'C' {
-							InfoAddSwapCached(t.Builder, int64(n))
+							InfoAddSwapCached(t.Builder, n)
 							continue
 						}
 						if t.Val[4] == 'F' {
-							InfoAddSwapFree(t.Builder, int64(n))
+							InfoAddSwapFree(t.Builder, n)
 							continue
 						}
-						InfoAddSwapTotal(t.Builder, int64(n))
+						InfoAddSwapTotal(t.Builder, n)
 						continue
 					}
 					if v == 'h' {
-						InfoAddShmem(t.Builder, int64(n))
+						InfoAddShmem(t.Builder, n)
 						continue
 					}
 					if v == 'l' {
-						InfoAddSlab(t.Builder, int64(n))
+						InfoAddSlab(t.Builder, n)
 						continue
 					}
 					if v == 'R' {
-						InfoAddSReclaimable(t.Builder, int64(n))
+						InfoAddSReclaimable(t.Builder, n)
 						continue
 					}
-					InfoAddSUnreclaim(t.Builder, int64(n))
+					InfoAddSUnreclaim(t.Builder, n)
 					continue
 				}
 				if v == 'V' {
 					if t.Val[8] == 'C' {
-						InfoAddVmallocChunk(t.Builder, int64(n))
+						InfoAddVmallocChunk(t.Builder, n)
 						continue
 					}
 					if t.Val[8] == 'T' {
-						InfoAddVmallocTotal(t.Builder, int64(n))
+						InfoAddVmallocTotal(t.Builder, n)
 						continue
 					}
-					InfoAddVmallocUsed(t.Builder, int64(n))
+					InfoAddVmallocUsed(t.Builder, n)
 					continue
 				}
 				if v == 'W' {
 					if nameLen == 9 {
-						InfoAddWriteback(t.Builder, int64(n))
+						InfoAddWriteback(t.Builder, n)
 						continue
 					}
-					InfoAddWritebackTmp(t.Builder, int64(n))
+					InfoAddWritebackTmp(t.Builder, n)
 					continue
 				}
 				if v == 'B' {
 					if nameLen == 6 {
-						InfoAddBounce(t.Builder, int64(n))
+						InfoAddBounce(t.Builder, n)
 						continue
 					}
-					InfoAddBuffers(t.Builder, int64(n))
+					InfoAddBuffers(t.Builder, n)
 					continue
 				}
 				if v == 'K' {
-					InfoAddKernelStack(t.Builder, int64(n))
+					InfoAddKernelStack(t.Builder, n)
 					continue
 				}
 				if v == 'N' {
-					InfoAddNFSUnstable(t.Builder, int64(n))
+					InfoAddNFSUnstable(t.Builder, n)
 					continue
 				}
 				if v == 'P' {
-					InfoAddPageTables(t.Builder, int64(n))
+					InfoAddPageTables(t.Builder, n)
 				}
-				InfoAddUnevictable(t.Builder, int64(n))
+				InfoAddUnevictable(t.Builder, n)
 			}
 			t.Builder.Finish(InfoEnd(t.Builder))
 			t.Data <- t.Profiler.Builder.Bytes[t.Builder.Head():]
